A lot of the memory Quelea used (about half) was down to images not being loaded and dealt with as efficiently as they could be. This has now hopefully been sorted with the introduction of an image manager, and making sure the preview images are only loaded into memory at the size they need to be.
Previously Quelea somewhat struggled when given 1200MB of RAM to work with, with these changes it comfortably fits in half that. Initial tests show it running somewhat acceptably when given 500MB of RAM, which never would have happened previously - though this is very much a lower limit and not something to be recommended using in practice!