Add dissolve transition for Video Output, but was able to keep existing system for slower systems. Created checkbox in options to enable advanced playback features (dissolving).
The Advanced Media Player is an out of process media player, meaning that it is two different
players to VLCJ and the system, as to not crash the primary JVM. log.txt and log2.txt log errors from these out of process video displays.


In addition created import option for CCLI's SongSelect, though it may not work outside of US.
(Would need to be tested by someone who would be outside of the US).